Implement an indexed glyph brush
In some cases we know exactly which sections we have and when we'll be
adding more or removing some sections. In such a case it's not worth
hashing, as it adds more complexity to the user - they need to re-queue
every section every time there's an update.

This patch introduces a new glyph-brush type that allows you to push and
pop sections by yourself, letting management of the indices be handled
by the user.