Implement a tagged glyph generator
This patch puts the generated vertices in a `Vec<Vec<_>>`, which is
useful because we want to be able to be able to distinguish vertices
from their respective text sections.

Why do this? Because when I render I do not query this library for every
frame. Instead, when a section is added, all `Section`s are `queue`d and
updated. Data from glyph-brush is then dumped into Vulkan buffers so
they can be redrawn efficiently.

Not only that, but having the text data in distinguishable buffer ranges
allows for various per-section and per-character transformations. For
instance, a single character of a section might be scaled or rotated or
translated differently from others. In addition to this, we can also
apply a transformation matrix to the entirety of the section's
associated UV data. I don't think it's proper for this library to
assume this role.

To add to this: I really would like to see an API where we don't need to
re-queue everything. My assumptions are that texels pretty much stay in
a DEVICE_LOCAL texture that is only updated when a very rare glyph
appears.

For games you'd likely just load every character that the font supports at
the beginning to avoid having to perform image updates while the process
is running. This is especially problematic in Vulkan as we'd need to do
the operation N times for N swapchain images, as we can't edit images
that are being used by frames that are in-flight. It's also fairly
expensive to keep streaming new texel data due to invalidation to the
GPU every so often.